App/requirements.txt
Flask
torch
torchvision
opencv-python
face_recognition
numpy
Pillow
